% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{as.matrix.backwardCriterion}
\alias{as.matrix.backwardCriterion}
\title{Convert \link[glmtoolbox.tzh]{backwardCriterion} to \link[base]{matrix}}
\usage{
\method{as.matrix}{backwardCriterion}(x, ...)
}
\arguments{
\item{x}{returned object from function \link[glmtoolbox.tzh]{backwardCriterion}.}

\item{...}{additional parameters, currently of no use}
}
\description{
Convert \link[glmtoolbox.tzh]{backwardCriterion} to \link[base]{matrix}
}
\keyword{internal}
