% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{as_flextable.backwardCriterion}
\alias{as_flextable.backwardCriterion}
\title{Convert \link[glmtoolbox.tzh]{backwardCriterion} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{backwardCriterion}(x, row.title = attr(z, which = "row.title", exact = TRUE), ...)
}
\arguments{
\item{x}{returned object from function \link[glmtoolbox.tzh]{backwardCriterion}.}

\item{...}{additional parameters, currently of no use}
}
\description{
Convert \link[glmtoolbox.tzh]{backwardCriterion} to \link[flextable]{flextable}
}
\keyword{internal}
