% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedData.R
\name{autoplot.groupedData}
\alias{autoplot.groupedData}
\title{\link[nlme]{groupedData} Objects}
\usage{
\method{autoplot}{groupedData}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{..}
}
\description{
..
}
\examples{
library(ggplot2)
autoplot(Indometh)
autoplot(Loblolly)
autoplot(ChickWeight) # Chick and Diet already factor

autoplot(nlme::Soybean) # two outer variables
autoplot(nlme::Glucose2) # nested groups

}
