% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{backwardCriterion}
\alias{backwardCriterion}
\title{\link{backwardCriterion}}
\usage{
backwardCriterion(model, trace = FALSE, ...)
}
\arguments{
\item{model, trace, ...}{parameters of function \link[glmtoolbox]{stepCriterion},
\strong{other than \code{direction}}}
}
\description{
\link{backwardCriterion}
}
\examples{
library(glmtoolbox)
# ?stepCriterion
data(depression)
m = glmgee(depressd ~ visit * group, id=subj, family=binomial(probit), 
 corstr="AR-M-dependent", data=depression)
list(
 'back' = m |> 
   backwardCriterion()
) |> fastmd::render2html()

}
\keyword{internal}
