% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{family.glmgee}
\alias{family.glmgee}
\title{\link[stats]{family} of \link[glmtoolbox]{glmgee} objects}
\usage{
\method{family}{glmgee}(object, ...)
}
\arguments{
\item{object}{\link[glmtoolbox]{glmgee} object}

\item{...}{S3 method dispatch holder}
}
\description{
\link[stats]{family} of \link[glmtoolbox]{glmgee} objects
}
