% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{glmgee}
\alias{glmgee}
\title{Fast Markdown Lines for \link[glmtoolbox]{glmgee}}
\description{
Fast Markdown Lines for \link[glmtoolbox]{glmgee}
}
\examples{
library(ecip); list(
 '`glmgee`' = glmgee(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> fastmd::render2html()

}
