% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{glmgee_S3}
\alias{glmgee_S3}
\alias{coef_.glmgee}
\alias{.pval.glmgee}
\alias{confint_.glmgee}
\alias{nobsText.glmgee}
\alias{desc_.glmgee}
\title{Auxiliary S3 method dispatch for \link[glmtoolbox]{glmgee} objects}
\usage{
\method{coef_}{glmgee}(x)

\method{.pval}{glmgee}(x)

\method{confint_}{glmgee}(x, level = 0.95, ...)

\method{nobsText}{glmgee}(x)

\method{desc_}{glmgee}(x, ...)
}
\arguments{
\item{x}{\link[glmtoolbox]{glmgee} object}

\item{level}{..}

\item{...}{..}
}
\description{
Auxiliary S3 method dispatch for \link[glmtoolbox]{glmgee} objects
}
\examples{
library(ecip); list(
 '`glmgee`' = glmgee(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> fastmd::render2html()

}
\keyword{internal}
