% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{glmtoolbox_bib}
\alias{glmtoolbox_bib}
\alias{.liang_zeger86}
\title{bibs in \pkg{ranef.tzh}}
\usage{
.liang_zeger86(key = "LiangZeger86", ...)
}
\arguments{
\item{key, ...}{\link[utils]{bibentry}}
}
\description{
bibs in \pkg{ranef.tzh}
}
\keyword{internal}
