% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_gls.R
\name{lme_gls}
\alias{lme_gls}
\title{\link[nlme]{lmeObject} and \link[nlme]{glsObject}}
\description{
\link[nlme]{lmeObject} and \link[nlme]{glsObject}
}
\examples{
library(nlme)
lme(fixed = distance ~ Sex * I(age-11), 
  weights = varIdent(form = ~ 1 | Sex), data = Orthodont)

m1 = lme(distance ~ age, data = Orthodont, keep.data = TRUE)
m2 = gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), 
  data = Ovary, correlation = corAR1(form = ~ 1 | Mare)) 
  
library(ecip); list(
 '`lme`' = m1, 
 '`gls`' = m2
) |> fastmd::render2html()
}
