% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{md_.backwardCriterion}
\alias{md_.backwardCriterion}
\title{R Markdown Lines for \link{backwardCriterion}}
\usage{
\method{md_}{backwardCriterion}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link{backwardCriterion}
}
\keyword{internal}
