% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{md_.glmgee}
\alias{md_.glmgee}
\title{R Markdown Lines for \link[glmtoolbox]{glmgee} Object}
\usage{
\method{md_}{glmgee}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[glmtoolbox]{glmgee} Object
}
\examples{
library(rmd.tzh); library(ecip); list(
 '`glmgee`' = glmgee(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> render_(file = 'glmgee')

}
\keyword{internal}
