% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R
\name{merMod}
\alias{merMod}
\title{\link[lme4]{merMod} Objects}
\description{
..
}
\examples{
# see ?lme4::glmer or ?lme4::cbpp
library(lme4)
m1 = glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), 
  data = cbpp, family = binomial)

startvec = c(Asym = 200, xmid = 725, scal = 350)
m2 = nlmer(circumference ~ SSlogis(age, Asym, xmid, scal) ~ Asym|Tree,
  Orange, start = startvec)

#library(HSAUR3)
m3 = glmer(outcome ~ treatment + visit + (1|patientID), 
  data = toenail, family = binomial, nAGQ = 20)

library(ecip); list(
 '`glmerMod`' = list(m1, m3)#,
 # '`nlmerMod`' = m2 # currently error
) |> fastmd::render2html()
}
