% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ranef.tzh-package}
\alias{ranef.tzh}
\alias{ranef.tzh-package}
\title{ranef.tzh: ranef.tzh}
\description{
Tingting's extension of the packages 'nlme', 'lme4', 'glmtoolbox'.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
