% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmerMod.R
\name{rlmerMod}
\alias{rlmerMod}
\title{\link[robustlmm]{rlmer} Objects}
\description{
..
}
\note{
Package \CRANpkg{robustlmm} imports package \CRANpkg{lme4}.

Object returned by function \link[robustlmm]{rlmer} does \strong{not} \link[base]{inherits} from \link[lme4]{merMod} class.
}
\examples{
# ?robustlmm::rlmer
library(robustlmm)
m = rlmer(Reaction ~ Days + (Days|Subject), data = sleepstudy,
 rho.sigma.e = psi2propII(smoothPsi, k = 2.28),
 rho.b = chgDefaults(smoothPsi, k = 5.14, s=10),
 rho.sigma.b = chgDefaults(smoothPsi, k = 5.14, s=10))
library(ecip); list(
 '`rlmerMod`' = m
) |> fastmd::render2html()
}
