% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{Sprintf.backwardCriterion}
\alias{Sprintf.backwardCriterion}
\title{\code{\link[=Sprintf.backwardCriterion]{Sprintf.backwardCriterion()}}}
\usage{
\method{Sprintf}{backwardCriterion}(x)
}
\arguments{
\item{x}{..}
}
\description{
\code{\link[=Sprintf.backwardCriterion]{Sprintf.backwardCriterion()}}
}
\keyword{internal}
