% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{stepCriterion_}
\alias{stepCriterion_}
\title{\link[glmtoolbox]{stepCriterion} with returned class}
\usage{
stepCriterion_(model, ...)
}
\arguments{
\item{model, ...}{see function \link[glmtoolbox]{stepCriterion}}
}
\description{
\link[glmtoolbox]{stepCriterion} with returned class
}
