% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{textCriterion}
\alias{textCriterion}
\title{\link{textCriterion}}
\usage{
textCriterion(x)
}
\arguments{
\item{x}{returned object from function \link[glmtoolbox]{stepCriterion}}
}
\description{
\link{textCriterion}
}
\keyword{internal}
