% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepCriterion_.R
\name{backwardCriterion}
\alias{backwardCriterion}
\title{\link{backwardCriterion}}
\usage{
backwardCriterion(model, ...)
}
\arguments{
\item{model, ...}{parameters of function \link[glmtoolbox]{stepCriterion},
\strong{other than \code{direction}}}
}
\description{
\link{backwardCriterion}
}
