% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{glmgee}
\alias{glmgee}
\title{\link[glmtoolbox]{glmgee} Objects}
\description{
\link[glmtoolbox]{glmgee} Objects
}
\examples{
library(ecip); list(
 '`glmgee`' = glmgee(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> fastmd::render2html()

}
