% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmgee.R
\name{glmgee_S3}
\alias{glmgee_S3}
\alias{coef_.glmgee}
\alias{.pval.glmgee}
\alias{confint_.glmgee}
\alias{nobsText.glmgee}
\alias{desc_.glmgee}
\title{Auxiliary S3 method dispatch for \link[glmtoolbox]{glmgee} objects}
\usage{
coef_.glmgee(x)

.pval.glmgee(x)

confint_.glmgee(x, level = 0.95, ...)

nobsText.glmgee(x)

desc_.glmgee(x, ...)
}
\arguments{
\item{x}{\link[glmtoolbox]{glmgee} object}

\item{level}{..}

\item{...}{..}
}
\description{
Auxiliary S3 method dispatch for \link[glmtoolbox]{glmgee} objects
}
